Remove compiler warning
svn path=/trunk/kdebase/kcontrol/; revision=172044

+++ kcontrol/fonts/kxftconfig.cpp
@@ -46,7 +46,7 @@ static QString dirSyntax(const QString &d)
 
         int slashPos=ds.findRev('/');
 
-        if(slashPos!=(ds.length()-1))
+        if(slashPos!=(((int)ds.length())-1))
             ds.append('/');
 
         return ds;
